Split out StateWrite/StateRead
