+++ storage2/src/state.rs
@@ -1,11 +1,11 @@
-use std::fmt::Debug;
-
 use async_trait::async_trait;
 
-use penumbra_proto::{Message, Protobuf};
-
+mod read;
 mod transaction;
+mod write;
+pub use read::StateRead;
 pub use transaction::Transaction as StateTransaction;
+pub use write::StateWrite;
 
 /// State is a lightweight copy-on-write fork of the chain state,
 /// implemented as a RYW cache over a pinned JMT version.
@@ -39,70 +39,4 @@ impl StateRead for State {
     fn get_raw(&self, key: String) -> Option<Vec<u8>> {
         todo!()
     }
-}
-
-#[async_trait]
-pub trait StateRead {
-    /// Get
-    fn get_raw(&self, key: String) -> Option<Vec<u8>>;
-
-    /// Gets a domain type from the State.
-    fn get<D, P>(&self, key: String) -> Option<D>
-    where
-        D: Protobuf<P>,
-        // TODO: does this get less awful if P is an associated type of D?
-        P: Message + Default,
-        P: From<D>,
-        D: TryFrom<P> + Clone + Debug,
-        <D as TryFrom<P>>::Error: Into<anyhow::Error>,
-    {
-        todo!()
-    }
-
-    /// Gets a proto type from the State.
-    fn get_proto<D, P>(&self, key: String) -> Option<P>
-    where
-        D: Protobuf<P>,
-        // TODO: does this get less awful if P is an associated type of D?
-        P: Message + Default,
-        P: From<D>,
-        D: TryFrom<P> + Clone + Debug,
-        <D as TryFrom<P>>::Error: Into<anyhow::Error>,
-    {
-        todo!()
-    }
-}
-
-pub trait StateWrite {
-    /// Copy-on-write put
-    fn put_raw(&mut self, key: String, value: Vec<u8>);
-
-    /// Sets a domain type from the State.
-    fn put<D, P>(&self, key: String, value: D)
-    where
-        D: Protobuf<P>,
-        // TODO: does this get less awful if P is an associated type of D?
-        P: Message + Default,
-        P: From<D>,
-        D: TryFrom<P> + Clone + Debug,
-        <D as TryFrom<P>>::Error: Into<anyhow::Error>,
-    {
-        todo!()
-    }
-
-    /// Puts a proto type on the State.
-    fn put_proto<D, P>(&self, key: String, value: P)
-    where
-        D: Protobuf<P>,
-        // TODO: does this get less awful if P is an associated type of D?
-        P: Message + Default,
-        P: From<D>,
-        D: TryFrom<P> + Clone + Debug,
-        <D as TryFrom<P>>::Error: Into<anyhow::Error>,
-    {
-        todo!()
-    }
-
-    /// Delete a key from state.
-    fn delete(&mut self, key: String);
 }
